Refactor: refactored the entire background manager as well as the RP creation, allowing it to support both animated (using mcmeta files) and regular textures. The RP does not overwrite files anymore, just adds any new background you might have.

+++ common/src/main/java/com/stalemated/customtooltips/util/ResourcepackManager.java
@@ -2,6 +2,7 @@ package com.stalemated.customtooltips.util;
 
 import java.io.File;
 import java.io.FileWriter;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import com.stalemated.lib.helper.PlatformHelper;
@@ -14,26 +15,52 @@ public class ResourcepackManager {
         try {
             if (!Files.exists(RESOURCE_PACK_DIR)) Files.createDirectories(RESOURCE_PACK_DIR);
 
-            File mcmeta = RESOURCE_PACK_DIR.resolve("pack.mcmeta").toFile();
-            try (FileWriter writer = new FileWriter(mcmeta)) {
-                writer.write(getPackMcmetaString());
-            }
+            generatePackMcmetaFile();
 
             CustomFontManager.loadAndGenerateFonts();
             CustomBackgroundManager.loadAndGenerateBackgrounds();
+
+            generateAtlasJson();
         } catch (Exception e) {
             LOGGER.error("Failed to generate resource pack", e);
         }
     }
 
-    private static String getPackMcmetaString() {
-        return """
-                {
-                   "pack": {
-                       "pack_format": 15,
-                       "description": "Custom Tooltip API - Auto Generated Assets"
-                   }
-                }
-                """;
+    private static void generatePackMcmetaFile() throws IOException {
+        File mcmeta = RESOURCE_PACK_DIR.resolve("pack.mcmeta").toFile();
+        if (!mcmeta.exists()) {
+            try (FileWriter writer = new FileWriter(mcmeta)) {
+                writer.write("""
+                    {
+                        "pack": {
+                           "pack_format": 15,
+                           "description": "Custom Tooltip API - Auto Generated Assets"
+                        }
+                    }
+                    """);
+            }
+        }
+    }
+
+    private static void generateAtlasJson() throws Exception {
+        Path atlasesDir = RESOURCE_PACK_DIR.resolve("assets").resolve("minecraft").resolve("atlases");
+        if (!Files.exists(atlasesDir)) Files.createDirectories(atlasesDir);
+
+        File blocksJson = atlasesDir.resolve("blocks.json").toFile();
+        if (!blocksJson.exists()) {
+            try (FileWriter writer = new FileWriter(blocksJson)) {
+                writer.write("""
+                    {
+                        "sources": [
+                            {
+                                "type": "directory",
+                                "source": "gui/tooltip_backgrounds",
+                                "prefix": "gui/tooltip_backgrounds/"
+                            }
+                        ]
+                    }
+                    """);
+            }
+        }
     }
 }
